refactor: remove unnecessary browser distinction on data initialization
