--- src/utils/data.browser.ts
@@ -1,21 +1,0 @@
-import { Readable } from 'stream'
-
-// eslint-disable-next-line require-await
-export async function prepareData(data: string | Buffer | Readable): Promise<Buffer> {
-  // string
-  if (typeof data === 'string') return Buffer.from(data)
-
-  // buffer
-  if (Buffer.isBuffer(data)) return data
-
-  // readable
-  return new Promise(resolve => {
-    const buffers: Array<Uint8Array> = []
-    data.on('data', d => {
-      buffers.push(d)
-    })
-    data.on('end', () => {
-      resolve(Buffer.concat(buffers))
-    })
-  })
-}
